chore: introduced AuthController.ts and AuthService.ts which handle appropiate responsibilities
